+++ Assets/Mirage/Editor/NetworkScenePostProcess.cs
@@ -57,7 +57,7 @@ namespace Mirage
                     }
                     // throwing an exception would only show it for one object
                     // because this function would return afterwards.
-                    else logger.LogError("Scene " + identity.gameObject.scene.path + " needs to be opened and resaved, because the scene object " + identity.name + " has no valid sceneId yet.");
+                    else logger.LogWarning("Scene " + identity.gameObject.scene.path + " needs to be opened and resaved, because the scene object " + identity.name + " has no valid sceneId yet.");
                 }
             }
         }