+++ Assets/Mirage/Runtime/CharacterSpawner.cs
@@ -43,11 +43,11 @@ namespace Mirage
             {
                 if (SceneManager != null)
                 {
-                    SceneManager.ClientSceneChanged.AddListener(OnClientSceneChanged);
+                    SceneManager.OnClientFinishedSceneChange.AddListener(OnClientFinishedSceneChange);
                 }
                 else
                 {
-                    Client.Authenticated.AddListener(c => Client.Send(new AddCharacterMessage()));
+                    Client.Authenticated.AddListener(OnClientAuthenticated);
                 }
 
                 if (ClientObjectManager != null)
@@ -73,13 +73,18 @@ namespace Mirage
         {
             if (Client != null && SceneManager != null)
             {
-                SceneManager.ClientSceneChanged.RemoveListener(OnClientSceneChanged);
-                Client.Authenticated.RemoveListener(c => Client.Send(new AddCharacterMessage()));
+                SceneManager.OnClientFinishedSceneChange.RemoveListener(OnClientFinishedSceneChange);
+                Client.Authenticated.RemoveListener(OnClientAuthenticated);
             }
             if (Server != null)
             {
                 Server.Started.RemoveListener(OnServerStarted);
             }
+        }
+
+        private void OnClientAuthenticated(INetworkPlayer _)
+        {
+            Client.Send(new AddCharacterMessage());
         }
 
         private void OnServerStarted()
@@ -91,8 +96,9 @@ namespace Mirage
         /// Called on the client when a normal scene change happens.
         /// <para>The default implementation of this function sets the client as ready and adds a player. Override the function to dictate what happens when the client connects.</para>
         /// </summary>
-        /// <param name="conn">Connection to the server.</param>
-        private void OnClientSceneChanged(string sceneName, SceneOperation sceneOperation)
+        /// <param name="scenePath"></param>
+        /// <param name="sceneOperation">The type of scene load that happened.</param>
+        public virtual void OnClientFinishedSceneChange(string scenePath, SceneOperation sceneOperation)
         {
             if (AutoSpawn && sceneOperation == SceneOperation.Normal)
                 RequestServerSpawnPlayer();