+++ Assets/Mirage/Runtime/Messages.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace Mirage
@@ -18,10 +19,10 @@ namespace Mirage
     [NetworkMessage]
     public struct SceneMessage
     {
-        public string scenePath;
+        public string MainActivateScene;
         // Normal = 0, LoadAdditive = 1, UnloadAdditive = 2
-        public SceneOperation sceneOperation;
-        public string[] additiveScenes;
+        public SceneOperation SceneOperation;
+        public List<string> AdditiveScenes;
     }
 
     [NetworkMessage]