+++ Assets/Mirage/Runtime/SocketLayer/AssemblyInfo.cs
@@ -1,6 +1,6 @@
 using System.Reflection;
 using System.Runtime.CompilerServices;
 
-[assembly: AssemblyVersion("101.9.2")]
+[assembly: AssemblyVersion("102.0.0")]
 
 [assembly: InternalsVisibleTo("Mirage.SocketLayer.Tests")]