+++ Assets/Mirage/Samples~/RigidbodyPhysics/Scripts/AssemblyInfo.cs
@@ -1,3 +1,3 @@
 ﻿using System.Reflection;
 
-[assembly: AssemblyVersion("101.9.2")]
+[assembly: AssemblyVersion("102.0.0")]